test: check that postgresql is started before running the tests

+++ acceptance.sh
@@ -42,6 +42,9 @@ function skip()
 
 function phobos_setup()
 {
+    sudo systemctl status postgresql >/dev/null 2>&1||
+        error "postgresql service not started"
+    HAVE_POSTGRES=y
     sudo -u postgres phobos_db setup_db -p phobos
     phobos_db setup_tables
 
@@ -74,6 +77,11 @@ function phobos_setup()
 
 function phobos_teardown()
 {
+    if [ "$HAVE_POSTGRES" != "y" ]
+    then
+        return
+    fi
+
     kill "$PHOBOSD_PID"
     phobos_db drop_tables
     sudo -u postgres phobos_db drop_db
